requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.2
colorama==0.4.6
faiss-cpu==1.11.0
filelock==3.18.0
fsspec==2025.7.0
huggingface-hub==0.33.4
idna==3.10
-e git+https://github.com/erizosamurai/imgio.git@ed8be1c37b261f8c14431dff4523e6fc4557bb2e#egg=imagio
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
packaging==25.0
pillow==11.3.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==80.9.0
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.3
typing_extensions==4.14.1
urllib3==2.5.0
